refactor: constant common 패키지로 위치 변경
