+++ src/main/java/io/github/jeli01/kakao_bootcamp_community/board/dto/response/GetBoardsResponse.java
@@ -2,7 +2,7 @@ package io.github.jeli01.kakao_bootcamp_community.board.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.github.jeli01.kakao_bootcamp_community.board.dto.BoardsInnerData;
-import io.github.jeli01.kakao_bootcamp_community.util.contant.ErrorMessage;
+import io.github.jeli01.kakao_bootcamp_community.common.contant.ErrorMessage;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
